Implement recommend component
