+++ client/ratings_reviews/ReviewList.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { FormControl, Select, MenuItem, Grid, Button } from '@material-ui/core';
 import ReviewItem from './ReviewItem.jsx';
-import ReviewForm from './ReviewForm.jsx';
+import AddReview from './AddReview.jsx';
 
 const ReviewList = (props) => {
   const { list, id, displayBySort } = props;
@@ -68,7 +68,7 @@ const ReviewList = (props) => {
         <br />
         <Grid container direction='row'>
           {renderMoreReviewButton()}
-          <ReviewForm />
+          <AddReview />
         </Grid>
       </div>
     </div>
